Separate position calculation from item style
