+++ packages/react-native-sortables/src/utils/reanimated/index.ts
@@ -1,2 +1,3 @@
 export * from './animatedTimeout';
 export { useAnimatedDebounce } from './useAnimatedDebounce';
+export { default as useMutableValue } from './useMutableValue';
