+++ packages/react-native-sortables/src/utils/reanimated/useMutableValue.ts
@@ -1,0 +1,7 @@
+import { useState } from 'react';
+import { makeMutable } from 'react-native-reanimated';
+
+export default function useMutableValue<T>(initialValue: T) {
+  const [mutable] = useState(makeMutable(initialValue));
+  return mutable;
+}
